Feature/code modulization (#7)
* DB connection Success

* Change file name

* Making Directory 'Mask_dataset'

* Fixed Path-Address

* Failed to Stack DB image

* Revert "Failed to Stack DB image"

This reverts commit efc360a0280b73049379369ef090e3dc380172b8.

* Add download image function on api

* Bug Fixed

* add processing on api

* upload saved_model.h5

* Complete Code_Modulization
