+++ original_app/mask_dataset/test-set/off_mask/ReadMe.txt
@@ -1,0 +1,2 @@
+You should put some picture No wearing mask in this directory
+
