+++ original_app/mask_dataset/train-set/off_mask/ReadMe.txt
@@ -1,0 +1,1 @@
+You should put some picture No wearing mask in this directory
