+++ original_app/mask_dataset/train-set/on_mask/ReadMe.txt
@@ -1,0 +1,1 @@
+You should put some picture wearing mask in this directory
